+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Clients\CommandRunner added
 * Exceptions\E4xx_CommandFailed added
 * Exceptions\E4xx_CommandSucceeded added
+* Exceptions\E5xx_CommandRunnerException added
+* Exceptions\E5xx_CommandFailedToStart added
 * ProcessRunners\PopenProcessRunner added
 * Requirements\RequireCommandFailed added
 * Requirements\RequireCommandSucceeded added